+++ finagle-http2/src/main/scala/com/twitter/finagle/http2/LoggerPerFrameTypeLogger.scala
@@ -5,10 +5,13 @@ import io.netty.buffer.{ByteBuf, ByteBufUtil}
 import io.netty.channel.ChannelHandlerContext
 import io.netty.handler.codec.http2.{Http2Flags, Http2FrameLogger, Http2Headers, Http2Settings}
 import io.netty.handler.logging.LogLevel
-import io.netty.util.internal.StringUtil
 
 private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     extends Http2FrameLogger(LogLevel.TRACE) {
+
+  // This level is symmetric with the level in `ChannelSnooper` which jives with its intended
+  // use of being a debug tool and not something to enable in production.
+  private[this] val level = Level.INFO
 
   private[this] val prefix = loggerNamePrefix
   private[this] val rootFrameLogger = Logger.get(prefix)
@@ -31,10 +34,10 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     padding: Int,
     endStream: Boolean
   ): Unit = {
-    if (dataLogger.isLoggable(Level.TRACE)) {
+    if (dataLogger.isLoggable(level)) {
       dataLogger.log(
-        Level.TRACE,
-        f"$direction ${ctx.channel} DATA: streamId=$streamId%d, padding=$padding%d, endStream=$endStream%b, length=${data.readableBytes}%d, bytes=${toString(dataLogger, data)}"
+        level,
+        f"$direction ${ctx.channel} DATA: streamId=$streamId%d, padding=$padding%d, endStream=$endStream%b, length=${data.readableBytes}%d, bytes=${toHexString(data)}"
       )
     }
   }
@@ -47,9 +50,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     padding: Int,
     endStream: Boolean
   ): Unit = {
-    if (headersLogger.isLoggable(Level.TRACE)) {
+    if (headersLogger.isLoggable(level)) {
       headersLogger.log(
-        Level.TRACE,
+        level,
         f"$direction ${ctx.channel} HEADERS: streamId=$streamId%d, headers=$headers, padding=$padding%d, endStream=$endStream%b"
       )
     }
@@ -66,9 +69,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     padding: Int,
     endStream: Boolean
   ): Unit = {
-    if (headersLogger.isLoggable(Level.TRACE)) {
+    if (headersLogger.isLoggable(level)) {
       headersLogger.log(
-        Level.TRACE,
+        level,
         f"$direction ${ctx.channel} HEADERS: streamId=$streamId%d, headers=$headers, streamDependency=$streamDependency%d, weight=$weight%d, exclusive=$exclusive%b, padding=$padding%d, endStream=$endStream%b"
       )
     }
@@ -82,9 +85,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     weight: Short,
     exclusive: Boolean
   ): Unit = {
-    if (priorityLogger.isLoggable(Level.TRACE)) {
+    if (priorityLogger.isLoggable(level)) {
       priorityLogger.log(
-        Level.TRACE,
+        level,
         f"$direction ${ctx.channel} PRIORITY: streamId=$streamId%d, streamDependency=$streamDependency%d, weight=$weight%d, exclusive=$exclusive%b"
       )
     }
@@ -96,9 +99,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     streamId: Int,
     errorCode: Long
   ): Unit = {
-    if (rstLogger.isLoggable(Level.TRACE)) {
+    if (rstLogger.isLoggable(level)) {
       rstLogger.log(
-        Level.TRACE,
+        level,
         f"$direction ${ctx.channel} RST_STREAM: streamId=$streamId%d, errorCode=$errorCode%d"
       )
     }
@@ -108,8 +111,8 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     direction: Http2FrameLogger.Direction,
     ctx: ChannelHandlerContext
   ): Unit = {
-    if (settingsLogger.isLoggable(Level.TRACE)) {
-      settingsLogger.log(Level.TRACE, s"$direction ${ctx.channel} SETTINGS: ack=true")
+    if (settingsLogger.isLoggable(level)) {
+      settingsLogger.log(level, s"$direction ${ctx.channel} SETTINGS: ack=true")
     }
   }
 
@@ -118,9 +121,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     ctx: ChannelHandlerContext,
     settings: Http2Settings
   ): Unit = {
-    if (settingsLogger.isLoggable(Level.TRACE)) {
+    if (settingsLogger.isLoggable(level)) {
       settingsLogger.log(
-        Level.TRACE,
+        level,
         s"$direction ${ctx.channel} SETTINGS: ack=false, settings=$settings"
       )
     }
@@ -131,9 +134,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     ctx: ChannelHandlerContext,
     data: Long
   ): Unit = {
-    if (pingLogger.isLoggable(Level.TRACE)) {
+    if (pingLogger.isLoggable(level)) {
       pingLogger.log(
-        Level.TRACE,
+        level,
         f"$direction ${ctx.channel} PING: ack=false, data=$data"
       )
     }
@@ -144,9 +147,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     ctx: ChannelHandlerContext,
     data: Long
   ): Unit = {
-    if (pingLogger.isLoggable(Level.TRACE)) {
+    if (pingLogger.isLoggable(level)) {
       pingLogger.log(
-        Level.TRACE,
+        level,
         f"$direction ${ctx.channel} PING: ack=true, data=$data"
       )
     }
@@ -160,9 +163,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     headers: Http2Headers,
     padding: Int
   ): Unit = {
-    if (pushPromiseLogger.isLoggable(Level.TRACE)) {
+    if (pushPromiseLogger.isLoggable(level)) {
       pushPromiseLogger.log(
-        Level.TRACE,
+        level,
         f"$direction ${ctx.channel} PUSH_PROMISE: streamId=$streamId%d, promisedStreamId=$promisedStreamId%d, headers=$headers, padding=$padding%d"
       )
     }
@@ -175,11 +178,10 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     errorCode: Long,
     debugData: ByteBuf
   ): Unit = {
-    if (goAwayLogger.isLoggable(Level.TRACE)) {
+    if (goAwayLogger.isLoggable(level)) {
       goAwayLogger.log(
-        Level.TRACE,
-        f"$direction ${ctx.channel} GO_AWAY: lastStreamId=$lastStreamId%d, errorCode=$errorCode%d, length=${debugData.readableBytes}%d, bytes=${this
-          .toString(goAwayLogger, debugData)}"
+        level,
+        f"$direction ${ctx.channel} GO_AWAY: lastStreamId=$lastStreamId%d, errorCode=$errorCode%d, length=${debugData.readableBytes}%d, bytes=${toHexString(debugData)}"
       )
     }
   }
@@ -190,9 +192,9 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     streamId: Int,
     windowSizeIncrement: Int
   ): Unit = {
-    if (windowUpdateLogger.isLoggable(Level.TRACE)) {
+    if (windowUpdateLogger.isLoggable(level)) {
       windowUpdateLogger.log(
-        Level.TRACE,
+        level,
         f"$direction ${ctx.channel} WINDOW_UPDATE: streamId=$streamId%d, windowSizeIncrement=$windowSizeIncrement%d"
       )
     }
@@ -206,20 +208,13 @@ private[http2] class LoggerPerFrameTypeLogger(loggerNamePrefix: String)
     flags: Http2Flags,
     data: ByteBuf
   ): Unit = {
-    if (unknownFrameLogger.isLoggable(Level.TRACE)) {
+    if (unknownFrameLogger.isLoggable(level)) {
       unknownFrameLogger.log(
-        Level.TRACE,
-        f"$direction ${ctx.channel} UNKNOWN: frameType=${frameType & 255}%d, streamId=$streamId%d, flags=${flags.value}%d, length=${data.readableBytes}%d, bytes=${this
-          .toString(unknownFrameLogger, data)}"
+        level,
+        f"$direction ${ctx.channel} UNKNOWN: frameType=${frameType & 255}%d, streamId=$streamId%d, flags=${flags.value}%d, length=${data.readableBytes}%d, bytes=${toHexString(data)}"
       )
     }
   }
 
-  private[this] def toString(logger: Logger, buf: ByteBuf): String = {
-    if (logger.isLoggable(Level.TRACE)) {
-      ByteBufUtil.hexDump(buf)
-    } else {
-      StringUtil.EMPTY_STRING
-    }
-  }
+  private[this] def toHexString(buf: ByteBuf): String = ByteBufUtil.hexDump(buf)
 }